feat: production-ready integration, monetisation & contract fixes
Bridge critical contract mismatch between Managing Director and Factory:
- Add brief_adapter.py: converts MD Pydantic BuildBrief → Factory v1 schema
- Add factory_callback.py: authenticated POST to /factory/callback with X-Factory-Secret
- Update factory-build.yml: add correlation_id/callback_url inputs, adapter + callback steps
- Update integration contract docs with callback payload and MD-format brief support

Monetisation readiness:
- Fix lead capture API to persist emails (JSONL file storage)
- Add GET /api/lead endpoint for lead retrieval
- Works on Vercel (/tmp) and local (./data)

Production hardening:
- Fix build_control.py business-score type (int→float) that blocked pipeline
- Add validate_env.py: preflight config validation
- Add requirements.txt and Dockerfile for containerised execution
- Add .dockerignore
- Add 5 cross-repo integration tests (run_integration_tests.py)
- Add MD-format test brief for integration testing
- Update CI to run integration tests
- Update README with Docker, callback, and testing docs

Verified: 11/11 factory tests pass, 5/5 integration tests pass,
full dry-run pipeline completes (17 steps, quality 12/12 PROCEED).

+++ scripts/build_control.py
@@ -122,7 +122,7 @@ def main() -> None:
     parser = argparse.ArgumentParser(description="Build control and rate limiting")
     parser.add_argument("--brief-file", required=True, help="Path to normalized brief JSON")
     parser.add_argument("--state-db", required=True, help="Path to lifecycle SQLite DB")
-    parser.add_argument("--business-score", type=int, default=0, help="Business gate score")
+    parser.add_argument("--business-score", type=float, default=0, help="Business gate score")
     parser.add_argument("--result-file", default="", help="Path to write control result JSON")
     parser.add_argument("--project-id", default="", help="Project identifier for logging")
     parser.add_argument("--dry-run", action="store_true", help="Check limits without blocking")
